Switch mnist dataset mirror to a more reliable one
The current mnist dataset mirror used in the cleverhans_v3 sometimes sporadically return 503. This issue was first noted in 2018 (#496) and considered resolved since everything seem to be working fine since then.
Yesterday, @orlp reported this issue again. I looked into it and it's quite sporadic, returning a 503 every once in a while. I thought it best to switch the url once and for all.
Since we're using the "https://storage.googleapis.com/cvdf-datasets/mnist/" in our jax tutorial, I wanted to stay consistent with that and changed all referenced " http://yann.lecun.com/exdb/mnist/" to "https://storage.googleapis.com/cvdf-datasets/mnist/". This is the same url referenced in the tf datasets class.